Pass `rm()` as function parameter, add more tests
Instead of passing the self-destruct function as `this`, this commit
changes the behavior so that it's passed as a secondary function
parameter.

This also adds a test to ensure that if you register two
listeners and the first one removes itself, the second one still gets
called without any problems.

> if you register two listeners and the first one removes itself, does the second one still get called: